Decompose the two worst cognitive-complexity alert formatters
TrackingDifferenceNotifier.formatBreach, at complexity 44 against the
threshold of 15, becomes the field-backed BreachMessageFormatter class,
since plain method extraction would have tipped the notifier into a god
class. LimitCheckNotifier.notify splits into per-section appenders with
an all-clear guard. Three pure helpers became static after the
extraction removed the transitive edge that had connected them to the
notifier's field-touching cluster. Existing tests already covered every
reshaped branch and ran unchanged.

Cognitive complexity violations 20 -> 18, PMD violations 128 -> 123
including three bonus cyclomatic-complexity eliminations.
