+++ metrics/scorecard.json
@@ -1,8 +1,8 @@
 {
-  "branchCoverage": 87.68,
+  "branchCoverage": 87.69,
   "classCouplingViolations": 34,
   "clockViolations": 0,
-  "cognitiveComplexityViolations": 20,
+  "cognitiveComplexityViolations": 18,
   "compilerWarnings": 0,
   "deepNestingViolations": 0,
   "disabledTests": 5,
@@ -11,7 +11,7 @@
   "godClasses": 2,
   "interactionVerifyingTestClasses": 363,
   "interactionVerifyingTestShare": 36.0,
-  "lineCoverage": 84.04,
+  "lineCoverage": 84.06,
   "longClasses": 0,
   "meanInstability": 0.473,
   "mocklessTestClasses": 465,
@@ -26,18 +26,18 @@
   "overMockedTestClasses": 113,
   "overMockedTestShare": 11.2,
   "pitestSlicedModules": 52,
-  "pmdViolations": 128,
+  "pmdViolations": 123,
   "pmdViolationsByRule": {
-    "CognitiveComplexity": 20,
+    "CognitiveComplexity": 18,
     "CouplingBetweenObjects": 34,
-    "CyclomaticComplexity": 69,
+    "CyclomaticComplexity": 66,
     "GodClass": 2,
     "TooManyMethods": 3
   },
-  "slowest10TestClassesSeconds": 119.4,
+  "slowest10TestClassesSeconds": 114.7,
   "springBootTests": 76,
-  "suiteWallTimeSeconds": 207,
-  "testClassesOver10Seconds": 4,
+  "suiteWallTimeSeconds": 206,
+  "testClassesOver10Seconds": 5,
   "testClassesTotal": 1008,
   "testIntegrationShare": 6.6,
   "testSliceShare": 13.7,
